# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /speaking/reset:
    post:
      operationId: resetSpeakingCourse
      tags: [ speaking ]
      summary: Wipe all student progress while preserving course content
      responses:
        "200":
          description: Course reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetCourseResult"
        "502":
          description: Reset failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /speaking/lectures/generate:
    post:
      operationId: generateSpeakingLectures
      tags: [ speaking ]
      summary: Author additional medium + long lecture variants for each unit
      responses:
        "200":
          description: Lectures generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateLecturesResult"
        "502":
          description: Lecture generation upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /speaking/narrative-report:
    post:
      operationId: generateNarrativeReport
      tags: [ speaking ]
      summary: Generate an AI written progress report from all logged work
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NarrativeReport"
        "502":
          description: Report generation upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required: [ error ]
      properties:
        error: { type: string }
...
    SpeakingLectureRef:
      type: object
      required: [ id, code, title ]
      properties:
        id: { type: integer }
        code: { type: string }
        title: { type: string }
...
    ResetCourseResult:
      type: object
      required: [ ok, deleted ]
      properties:
        ok: { type: boolean }
        deleted:
          type: object
          required:
            [ attempts, responses, activity, assessments, practiceAssignments, recordings ]
          properties:
            attempts: { type: integer }
            responses: { type: integer }
            activity: { type: integer }
            assessments: { type: integer }
            practiceAssignments: { type: integer }
            recordings: { type: integer }
    GenerateLecturesResult:
      type: object
      required: [ ok, created ]
      properties:
        ok: { type: boolean }
        created: { type: integer }
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/SpeakingLectureRef"
...
    NarrativeReport:
      type: object
      required: [ report, generatedAt ]
      properties:
        report: { type: string }
        generatedAt: { type: string, format: date-time }
...