# lib/api-zod/src/generated/types/generateLecturesResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SpeakingLectureRef } from './speakingLectureRef';

export interface GenerateLecturesResult {
  ok: boolean;
  created: number;
  lectures?: SpeakingLectureRef[];
}