# lib/api-zod/src/generated/types/resetCourseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResetCourseResultDeleted } from './resetCourseResultDeleted';

export interface ResetCourseResult {
  ok: boolean;
  deleted: ResetCourseResultDeleted;
}